update app.py and init.py
