+++ Week5/api/app.py
@@ -1,6 +1,26 @@
-from . import create_app
+from flask import Flask
 
-app = create_app()
+from .routes.inventory import inventory_bp
+
+
+def create_app():
+    """
+    Creates a Flask application.
+
+    This function initializes a Flask app and registers the API blueprint
+    to it. The logger is also set up to log errors to a file.
+
+    Returns:
+        app: The Flask application instance.
+    """
+    from Week3.utils import setup_logger
+
+    setup_logger()
+
+    app = Flask(__name__)
+    app.register_blueprint(inventory_bp, url_prefix="/api")
+    return app
+
 
 if __name__ == "__main__":
-    app.run(debug=True)
+    create_app().run(debug=True)
